requirements.txt
numpy==1.24.2
opencv-python-headless
Pillow==10.4.0
streamlit==1.36.0
tensorflow==2.15.0
